test: Move FP conversion tests to a separate file
