+++ test/unittests/CMakeLists.txt
@@ -19,6 +19,7 @@ target_sources(
     execute_call_test.cpp
     execute_control_test.cpp
     execute_death_test.cpp
+    execute_floating_point_conversion_test.cpp
     execute_floating_point_test.cpp
     execute_floating_point_test.hpp
     execute_numeric_test.cpp
